+++ sdk/internal/recording/example_test.go
@@ -26,8 +26,8 @@ func ExampleStop() {
 	}
 }
 
-func ExampleResetSanitizers() {
-	err := recording.ResetSanitizers(nil)
+func ExampleResetProxy() {
+	err := recording.ResetProxy(nil)
 	if err != nil {
 		panic(err)
 	}
@@ -119,7 +119,7 @@ func ExampleStart_second() {
 		// Test functionality
 
 		// Reset Sanitizers IF only for a single session
-		err = recording.ResetSanitizers(nil)
+		err = recording.ResetProxy(nil)
 		if err != nil {
 			panic(err)
 		}